[LNT] Python 3 support: use int() instead of long()
Use int() rather than long() to interpret a string as a long integer
since Python 3's integer are long by default. This was produced by
running futurize's stage2 lib2to3.fixes.fix_long.

Reviewers: cmatthews, hubert.reinterpretcast, kristof.beyls

Reviewed By: hubert.reinterpretcast

Subscribers: llvm-commits

Differential Revision: https://reviews.llvm.org/D67816

llvm-svn: 372820

+++ lnt/tests/compile.py
@@ -529,7 +529,7 @@ def test_build(base_name, run_info, variables, project, build_config, num_jobs,
                            " ".join("'%s'" % arg for arg in cmd))
             result = subprocess.check_output(cmd).strip()
             if result != "fail":
-                bytes = long(result)
+                bytes = int(result)
                 success = True
 
                 # For now, the way the software is set up things are going to
